Added `jacobi_driver.py` for directly testing Jacobi iteration.
Also documented it (it's probably permanent).
