+++ README.md
@@ -28,6 +28,13 @@ TODO: Add instructions on running these parts of the code.
 
 ## (d) `jacobi_iter`, `gs_iter`: Iteration Methods
 
+The actual "math" part of these functions can be found in `iterative_methods.py`.
+
+To test `jacobi_iter`, run:
+
+    python jacobi_driver.py <path_to_augmented_matrix> <path_to_x_0> <e> <m>
+    python jacobi_driver.py test/iter_aug.dat test/iter_x_0.dat 0.0001 10
+
 TODO: Add instructions on running these parts of the code.
 
 ## (e) `power_method`: Power Method
